+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Use `unitMultiplier` to calculate the price.
+- Add utmi campaign to the simulation request
+
 ## [2.144.0] - 2021-08-31
 ### Added
 - `priceDefinition` property to `OrderFormItem` type  on OrderForm.graphql.